requirements.txt
# Python package dependencies for the Telegram Summarizer Bot
# Run 'pip install -r requirements.txt'

# Core Telegram library with extensions
python-telegram-bot[ext]==21.2

# Google Generative AI library
google-generativeai==0.8.5

# For loading environment variables from .env file
python-dotenv==1.0.1

# Language detection
langdetect==1.0.9

# HTTP client used by underlying libraries
httpx==0.27.0

# Additional API error handling
google-api-core==2.17.1

# System utilities
psutil==5.9.6  # For monitoring memory usage

# Web server for webhook implementation
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Added dependencies for production use
pydantic>=2.0.0,<3.0.0
starlette>=0.27.0,<0.28.0